+++ src/webapp/src/flask.js
@@ -12,6 +12,8 @@ module.exports = function() {
       dbtDocsUrl: process.env.DBT_DOCS_URL,
       isSendAnonymousUsageStats: false,
       projectId: 'none',
+      // Update if we make a TOS update and require another acknowledgement for each UI user
+      termsOfServiceVersion: 1.0,
       version: 'source'
     }
   )